pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "s_spatioloji"
version = "0.1.0"
description = "scalable spatioloji"
readme = "README.md"
authors = [
  {name = "gynecoloji", email = "gynecoloji@gmail.com"}
]
maintainers = [
  {name = "gynecoloji", email = "gynecoloji@gmail.com"}
]
classifiers = [
    "Typing :: Typed",
    # TODO: Add additional classifiers as needed
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "rich",
  "zarr>=2.18,<3",
  "dask[array,dataframe]>=2024.1",
  "numpy>=1.26",
  "pandas>=2.0",
  "pyarrow>=14.0",
  "tifffile>=2024.1",
  "geopandas>=1.0",
  "shapely>=2.0",
  "scipy>=1.12",
  "scikit-learn>=1.4",
]

[project.optional-dependencies]
reduction  = ["umap-learn", "openTSNE"]
clustering = ["leidenalg", "igraph", "python-louvain"]
batch      = ["harmonypy", "pycombat>=0.3"]
imputation = ["scvi-tools>=1.0", "magic-impute", "rpy2"]
all        = ["s_spatioloji[reduction,clustering,batch,imputation]"]
requires-python = ">= 3.12"

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    { include-group = "typecheck" },
]
lint = [
    "ruff",
]
test = [
    "coverage",
    "pytest",
]
typecheck = [
    "ty",
]
docs = [
    "zensical",
    "mkdocstrings-python",
]

[project.urls]
bugs = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.project_slug }}/issues"
changelog = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.project_slug }}/releases"
documentation = "https://gynecoloji.github.io/s_spatioloji/"
homepage = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.project_slug }}"

[project.scripts]
s_spatioloji = "s_spatioloji.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.coverage.run]
branch = true
parallel = true
source = ["src/", "tests/"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
exclude_also = [
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@overload",
    "@typing.overload",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "\\.\\.\\.",
]

[tool.uv]
package = true
default-groups = ["dev"]
